backend/requirements.txt
blinker==1.8.2
cassandra-driver==3.29.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
geomet==0.2.1.post1
gevent==24.2.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python==2.0.2
MarkupSafe==2.1.5
pycparser==2.22
redis==5.1.1
setuptools==75.1.0
six==1.16.0
Werkzeug==3.0.4
zope.event==5.0
zope.interface==7.0.3
